Rename dialog to utils, move update checker
The dialog module is now named utils, and also now contains the
application update checking function.

In the future, generally useful but or "orphan" functions with no
obvious place to put them will end up here.
