+++ src/gui/install.rs
@@ -148,50 +148,3 @@ fn copy_dir_recursive(to: &PathBuf, from: &PathBuf) -> io::Result<()> {
 
   Ok(())
 }
-
-pub async fn get_master_version(local: ModVersionMeta) -> (String, Result<Option<ModVersionMeta>, String>) {
-  let res = send_request(local.remote_url.clone()).await;
-
-  match res {
-    Err(err) => (local.id, Err(err)),
-    Ok(remote) => {
-      if_chain! {
-        let mut stripped = String::new();
-        if strip_comments(remote.as_bytes()).read_to_string(&mut stripped).is_ok();
-        if let Ok(normalized) = handwritten_json::normalize(&stripped);
-        if let Ok(remote) = json5::from_str::<ModVersionMeta>(&normalized);
-        then {
-          if remote.version > local.version {
-            (
-              local.id,
-              Ok(Some(remote))
-            )
-          } else {
-            (
-              local.id,
-              Ok(None)
-            )
-          }
-        } else {
-          (
-            local.id,
-            Err(format!("Parse error. Payload:\n{}", remote))
-          )
-        }
-      }
-    }
-  }
-
-
-}
-
-async fn send_request(url: String) -> Result<String, String>{
-  reqwest::get(url)
-    .await
-    .map_err(|e| format!("{:?}", e))?
-    .error_for_status()
-    .map_err(|e| format!("{:?}", e))?
-    .text()
-    .await
-    .map_err(|e| format!("{:?}", e))
-}
